# Conflicts:
#	app/src/main/java/com/example/concordia_campus_guide/Activities/MainActivity.java
#	app/src/main/java/com/example/concordia_campus_guide/Database/Daos/RoomDao.java
#	app/src/main/java/com/example/concordia_campus_guide/Fragments/LocationFragment/LocationFragmentViewModel.java
#	app/src/main/java/com/example/concordia_campus_guide/Models/RoomModel.java
#	app/src/main/java/com/example/concordia_campus_guide/Models/Rooms.java

+++ app/src/main/java/com/example/concordia_campus_guide/Activities/MainActivity.java
@@ -1,6 +1,7 @@
 package com.example.concordia_campus_guide.Activities;
 
 import android.content.Intent;
+import android.location.Location;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -15,8 +16,9 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.concordia_campus_guide.Database.AppDatabase;
 import com.example.concordia_campus_guide.Fragments.InfoCardFragment.InfoCardFragment;
+import com.example.concordia_campus_guide.Fragments.LocationFragment.LocationFragment;
 import com.example.concordia_campus_guide.Global.ApplicationState;
-import com.example.concordia_campus_guide.Models.Building;
+import com.example.concordia_campus_guide.Global.SelectingToFromState;
 import com.example.concordia_campus_guide.Models.Buildings;
 import com.example.concordia_campus_guide.Models.Floors;
 import com.example.concordia_campus_guide.Models.Relations.BuildingWithFloors;
@@ -26,13 +28,12 @@ import com.example.concordia_campus_guide.Models.Rooms;
 import com.example.concordia_campus_guide.R;
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 
-import java.util.List;
-
 public class MainActivity extends AppCompatActivity {
 
     FragmentTransaction fragmentTransaction;
     FragmentManager fragmentManager;
     InfoCardFragment infoCardFragment;
+    LocationFragment locationFragment;
     MainActivityViewModel mViewModel;
     private BottomSheetBehavior swipeableInfoCard;
 
@@ -50,14 +51,9 @@ public class MainActivity extends AppCompatActivity {
         View infoCard = findViewById(R.id.info_card);
         swipeableInfoCard = BottomSheetBehavior.from(infoCard);
 
-        setUpDb();
+        locationFragment = (LocationFragment) getSupportFragmentManager().findFragmentById(R.id.locationFragment);
 
-        //examples on how to get objects from db
-        //begin
-        AppDatabase appDB = AppDatabase.getInstance(this);
-        List<Building> listBuildings = appDB.buildingDao().getAll();
-        List<BuildingWithFloors> buildingWithFloorsList = appDB.buildingDao().getBuildingsWithFloors();
-        List<FloorWithRooms> floorWithRooms = appDB.floorDao().getFloorWithRooms();
+        setUpDb();
     }
 
     @Override
@@ -70,10 +66,11 @@ public class MainActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
 
-        if (id == R.id.search) {
-            //TODO onClick of search_activity button
+        if(id == R.id.search){
             Intent openSearch = new Intent(MainActivity.this,
                     SearchActivity.class);
+
+            SelectingToFromState.setMyCurrentLocation(getMyCurrentLocation());
 
             startActivity(openSearch);
             return false;
@@ -94,7 +91,6 @@ public class MainActivity extends AppCompatActivity {
 
         infoCardFragment = new InfoCardFragment();
         infoCardFragment.setBuildingCode(buildingCode);
-
         fragmentTransaction = fragmentManager.beginTransaction();
         fragmentTransaction.add(R.id.info_card_frame, infoCardFragment);
         fragmentTransaction.commit();
@@ -125,21 +121,29 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    private void setUpDb() {
-        //delete previous db
-        getApplication().getApplicationContext().deleteDatabase(AppDatabase.DB_NAME);
+    private void setUpDb(){
+        if(!ApplicationState.getInstance(this).isDbIsSet()){
+            //delete previous db
+            getApplication().getApplicationContext().deleteDatabase(AppDatabase.DB_NAME);
 
-        //load buildings
-        Buildings buildings = ApplicationState.getInstance(this).getBuildings();
-        AppDatabase appDb = AppDatabase.getInstance(this);
-        appDb.buildingDao().insertAll(buildings.getBuildings());
+            //load buildings
+            Buildings buildings = ApplicationState.getInstance(this).getBuildings();
+            AppDatabase appDb = AppDatabase.getInstance(this);
+            appDb.buildingDao().insertAll(buildings.getBuildings());
 
-        //load floors
-        Floors floors = ApplicationState.getInstance(this).getFloors();
-        appDb.floorDao().insertAll(floors.getFloors());
+            //load floors
+            Floors floors = ApplicationState.getInstance(this).getFloors();
+            appDb.floorDao().insertAll(floors.getFloors());
 
-        //load rooms
-        Rooms rooms = ApplicationState.getInstance(this).getRooms();
-        appDb.roomDao().insertAll(rooms.getRooms());
+            //load rooms
+            Rooms rooms = ApplicationState.getInstance(this).getRooms();
+            appDb.roomDao().insertAll(rooms.getRooms());
+
+            ApplicationState.getInstance(this).setDbIsSetToTrue();
+        }
+    }
+
+    public Location getMyCurrentLocation(){
+        return this.locationFragment.getCurrentLocation();
     }
 }
